requirements.txt
streamlit
supabase
smartapi-python
pandas
logzero
pyotp
streamlit-lightweight-charts
requests
